requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
psycopg2-binary==2.9.9
python-decouple==3.8
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
uritemplate==4.1.1
